libtorrent: Optimize unsigned int datatypes
Every byte counts if we want to hit the CPU cache.

+++ libtorrent/src/net/throttle_internal.h
@@ -62,7 +62,7 @@ public:
 
 private:
   // Fraction is a fixed-precision value with the given number of bits after the decimal point.
-  static const uint32_t fraction_bits = 16;
+  static const uint8_t  fraction_bits = 16;
   static const uint32_t fraction_base = (1 << fraction_bits);
 
   typedef std::vector<ThrottleInternal*>  SlaveList;
